aws-infrastructure/requirements.txt
aws-cdk.core
aws-cdk.aws_codebuild
aws-cdk.aws_iam
aws-cdk.aws_codepipeline
aws-cdk.aws_codepipeline_actions
aws-cdk.aws_ec2
aws-cdk.aws_cloud9